+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ might include API breaking changes.
 Up for the next release
 -----------------------
 
+* Introduce `can_claim_draw()`, `can_claim_fifty_moves()` and
+  `can_claim_threefold_repitition()`.
+
 * Fix low level rotation operations `l90()`, `l45()` and `r45()`. There was
   no problem in core because correct versions of the functions were inlined.